Refactor expiration.py code
This commit changes the way we obtain expiration date
from remote host and port. Previous implementation
failed for any expired and incorrect certificate.
With current implementation we always get the number
of days, as we do not enforce certificate validity.
With this approach, we may analyse all certificate
properties in easier way.