Refactor checker
This commit updates the code of checker service.
First, a new Cert class with helper functions
to process the certificates of remote hosts
is introduced. Second, the function to match
hostnames with wildcard patterns is introduced.
Finally, the obsolete expiration.py is removed.
The details about scanned certificates are stored
in local database cache.